requirements.txt
powerline-status==2.7
setuptools==69.1.1
